Make gray action background full width in Android
